Reroute for latest AIO-SX AIO-DX install_deploy
I have modified the docs repo such that the landing pages
for original installation guides and deployment guides now
list the merged deploy_install_guides AIO-SX and AIO-DX
manuals instead of the older ones that are being replaced.

Links on the landing pages at this point for the original
controller_storage deployment guide and for the original
"blanket" installation manual still exist.  The reason I
left the one for the "blanket" installation guide is because
if someone clicks on that, there exists further links to
deployment information (guides) at the bottom.  Those links
have been modified to point to the AIO-SX and AIO-DX
deploy_install_guides.  The link for the controller_storage
deployment manual still goes to the original.

Once we get remaining combo manuals merged, we can further
modify these landing pages to where we eventually do not
have any manuals in the installation guides and
deployment guides area.

Change-Id: I7b436226c307b9bc803cc14d48712fb92c39975a
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>