Deploy Install Guides: AIO Duplex Extended
The All-In-One Duplex Extended deployment option extends the capacity
of the All-In-One Duplex deployment by adding up to four compute nodes
to the deployment.

The extended deployment option provides a capacity growth path for
someone starting with an All-In-One Duplex deployment.

Amended to add in comments from Greg Waines.

Depends-On: https://review.opendev.org/#/c/674799/

Story: 2005009
Task: 29485

Change-Id: I131fe72f7e298d6841f66ff7a6ccd3d197d7d2b1
Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>
Signed-off-by: Kristal Dale <kristal.dale@intel.com>