Release Note document for StarlingX 2.0
This is the initial draft of the release note document for stx.2.0.  I'd
like to thank Bill Zvonar for providing the initial input,  I have made
some formatting changes to his text to align with doc conventions.

There is an issue with the links to the storyboard entries.  They don't
work correctly and I don't know why.

- Patch: Removed trailing slash on links so that they would resolve
- Patch: Minor formatting and rewording based on feedback.

Change-Id: I02971dac25b2c3239a34319e2b07522ef6fea81b
Signed-off-by: Bruce Jones <bruce.e.jones@intel.com>
Signed-off-by: Kristal Dale <kristal.dale@intel.com>