Change Ironic Python Agent Ramdisk from CoreOS
The IPA CoreOS ramdisk has been deprecated for some time
and in discussing removing the out of date images, we identified
that there were still some references to it out there.

As such, this patch changes the documentation url to a more
appropriate image. Additionally, documentation for previously
released versions was not modified in order to maintain
user consistency.

Note: The choice of centos is ability to build and save the image.
      Nothing realistically prevents a debian based image from also
      being available, except job configuration. The ironic team
      would likely not object to such a job being proposed, but
      this is out of scope of just fixing an out of date URL
      in documentation.

Change-Id: Ifffbb9671979d41cf811be84f96e7236fd901d32