adjust yaml for cert-manager.io/v1alpha2 (ds, r6ds)
This commit should not be cherry-picked to r/stx.7.0, r/stx.8.0 or
starlingx/master.

The "organization" field in v1alpha2 API was moved from Certificate spec
to X509Subject spec in v1alpha3 and remains there in v1.  The new name
in X509Subject spec is "organizations".

r/stx.6.0 is using v1alpha2 API, so the "organization" field in Certificate
spec is used.

Reference:
https://cert-manager.io/v1.0-docs/reference/api-docs/#cert-manager.io%2fv1alpha2

Change-Id: I38229f2829ad0c2e31e735349cc22d877069216a
Signed-off-by: Michel Thebeau <Michel.Thebeau@windriver.com>