Change links provided in doc for users
In bullet point 4 inside the Host Setup topic, the link provided to
download the ISO gives the user too many options and unclear
instructions on wher exactly to find the ISO file. We propose that
the first link is always to the latest release and then additional
versions can be found on a second link. Other references to the
mirror are updated as well.

Additionally, on the Introduction page (https://docs.starlingx.io/),
the provided link is for release 8. We also changed it to point to
the latest release. The main page still shows the current version,
however now the change is pretty straighforwar for when new releases
come.

Closes-Bug: 2016032

Change-Id: I94201ef4f436b137f2601b033cb344d726384b55
Signed-off-by: Bruno Muniz <bruno.muniz@encora.com>