Updates "Automated Virtual Installation in StarlingX" documentation
Changed "Host Setup" to remove the command mkdir, described in item 6, to avoid creating the config folder as a subfolder to myconfig. This created a different path in relation to the one used in the script for the "Deployment Example".
Additionally, also in "Deployment Example", the file localhost.yml was changed to be prefixed by either simplex or duplex.

Closes-Bug: 2036446

Change-Id: Iceec865088fd0610e7c9c777e376666f69fe48d9